big changes to the ConvertBarFillAmountToTiles
I made huge tweaks to GraphicalBarConvertToTile.asm's ConvertBarFillAmountToTiles, instead of relying on gamemode $0100, to write which mode to $00, it now only relies on $00 itself, meaning you HAVE to write #$00 to $00 for level, layer 3, #$01 for level, sprite, and $02 for overworld, layer 3. This is simply because there are two versions for level: for layer 3, and for sprite, and the before-change version could not distinguish it.

Note that this will not apply to doublebar (seperate graphics) and ConvertBarFillAmountToTilesEdgeOverMultipleTiles. Because that would eat a TON of sprite tiles, and most hackers use the sprites more than the layer 3 status bar.