Image added showing displacement (not used yet)
Planning on adding a tutorial on how to place the graphical bar centered with the body of the sprite.

It's very possible that sprite's origin position may be in unexpected places, normally, it is somewhere on the top-left of its minimum bounding box, however, due to custom sprites may have their own handling, their origin XY pos may be something like this as to handle object collision.