BigIntFloorDiv now added
yeah, we want to make sure that even if the bar is centered, it also needs consistency reguardless of the bar's origin point being on the left or right side of the center. Truncate can result the bar being a pixel to the left or right when set to a different direction since it can round up and down.