sigh...
this is going to be a major update for the GMK boss. The fact that sprite graphic layouts drastically differs which means the tiles of the graphical bar have to be at different locations means that GraphicalBarConvertBarFillAmountToTilesIndirectAddressTable is necessary (graphical bar tile array are stored within sprite code, not the subroutine).