fixed a potential problem - using SA-1 ROM with SNESmathonly set to 1 breaks things
so I had to correct the defines to make sure uberasm test files and the routines uses the correct CPU and on the right registers. If you set snesmathonly to 1 on sa-1 rom, this will cause the routines to use snes math registers, while the test files calling the routines to still invoke sa-1, which causes problems.